crserver/requirements/common.txt
Django==4.1.1
djangorestframework==3.12.4
psycopg2==2.9.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.20.0
pytz==2023.3
celery==5.3.0
redis==4.4.2
django-celery-beat==2.5.0
flower==1.0.0



